refactor(public): move getAuctionHistory method
