+++ lib/public.js
@@ -1,20 +1,6 @@
 const request = require("request-promise");
 const Promise = require("bluebird");
-const {
-  API_LIMIT,
-  DEFAULT_SYMBOL,
-  DEFAULT_TIMEOUT,
-  EXCHANGE_API_URL,
-  SANDBOX_API_URL,
-  HEADERS,
-  _checkUndefined
-} = require("./utilities.js");
-
-/**
- * @callback Callback
- * @param error
- * @param data
- */
+const { HEADERS, _checkUndefined } = require("./utilities.js");
 
 class PublicClient {
   /**
@@ -67,31 +53,6 @@ class PublicClient {
       request(options)
         .then(data => resolve(data))
         .catch(error => reject(error.error || error));
-    });
-  }
-
-  /**
-   * @param {Object} [options={}]
-   * @param {string} [options.symbol] - Trading symbol.
-   * @param {number} [options.since] - Only return trades after this timestamp.
-   * @param {number} [options.limit_auction_results] - The maximum number of auction events to return.
-   * @param {boolean} [options.include_indicative] - Whether to include publication of indicative prices and quantities.
-   * @example
-   * const history = await publicClient.getAuctionHistory();
-   * @description Get the auction events.
-   * @see {@link https://docs.gemini.com/rest-api/#auction-history|auction-history}
-   */
-  getAuctionHistory({
-    symbol = this.symbol,
-    since,
-    limit_auction_results = API_LIMIT,
-    include_indicative
-  } = {}) {
-    return this.get({
-      uri: "v1/auction/" + symbol + "/history",
-      since,
-      limit_auction_results,
-      include_indicative
     });
   }
 }
